Change responsabilities of MessageDialog and MessageGraphicsRender
